chore: agent ui improvements
